UI improved in Favourites Fragment
