+++ app/src/main/java/org/fossasia/openevent/general/favorite/FavoriteFragment.kt
@@ -1,6 +1,5 @@
 package org.fossasia.openevent.general.favorite
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -12,14 +11,18 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import kotlinx.android.synthetic.main.fragment_favorite.*
-import kotlinx.android.synthetic.main.fragment_favorite.view.*
-import org.fossasia.openevent.general.MainActivity
+import kotlinx.android.synthetic.main.fragment_favorite.noLikedText
+import kotlinx.android.synthetic.main.fragment_favorite.view.favoriteEventsRecycler
+import kotlinx.android.synthetic.main.fragment_favorite.view.favoriteProgressBar
+import kotlinx.android.synthetic.main.fragment_favorite.view.currentDate
+import kotlinx.android.synthetic.main.fragment_favorite.view.searchToday
+import kotlinx.android.synthetic.main.fragment_favorite.view.searchTomorrow
+import kotlinx.android.synthetic.main.fragment_favorite.view.searchWeekend
+import kotlinx.android.synthetic.main.fragment_favorite.view.searchNextMonth
 import org.fossasia.openevent.general.R
 import org.fossasia.openevent.general.event.EVENT_ID
 import org.fossasia.openevent.general.event.Event
 import org.fossasia.openevent.general.event.EventDetailsFragment
-import org.fossasia.openevent.general.event.EventUtils
 import org.fossasia.openevent.general.event.FavoriteFabListener
 import org.fossasia.openevent.general.event.RecyclerViewClickListener
 import org.fossasia.openevent.general.search.SearchTimeViewModel
@@ -40,7 +43,6 @@ class FavoriteFragment : Fragment() {
     private val favoriteEventsRecyclerAdapter: FavoriteEventsRecyclerAdapter = FavoriteEventsRecyclerAdapter()
     private val favoriteEventViewModel by viewModel<FavouriteEventsViewModel>()
     private lateinit var rootView: View
-    private val TO_SEARCH: String = "ToSearchFragment"
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -84,41 +86,16 @@ class FavoriteFragment : Fragment() {
         val calendar = Calendar.getInstance()
         setCurrentDate(calendar)
         rootView.searchToday.setOnClickListener {
-            searchTimeViewModel.saveDate(EventUtils.getSimpleFormattedDate(calendar.time))
-            calendar.add(Calendar.DATE, 1)
-            searchTimeViewModel.saveNextDate(EventUtils.getSimpleFormattedDate(calendar.time))
-            searchTimeViewModel.saveDate(TODAY)
-            redirectToSearch()
+            favoriteEventViewModel.searchToday(calendar, searchTimeViewModel, activity)
         }
         rootView.searchTomorrow.setOnClickListener {
-            calendar.add(Calendar.DATE, 1)
-            searchTimeViewModel.saveNextDate(EventUtils.getSimpleFormattedDate(calendar.time))
-            calendar.add(Calendar.DATE, 1)
-            searchTimeViewModel.saveNextToNextDate(EventUtils.getSimpleFormattedDate(calendar.time))
-            searchTimeViewModel.saveDate(TOMORROW)
-            redirectToSearch()
+            favoriteEventViewModel.searchTomorrow(calendar, searchTimeViewModel, activity)
         }
         rootView.searchWeekend.setOnClickListener {
-            val today = calendar.get(Calendar.DAY_OF_WEEK)
-            if (today != Calendar.SATURDAY) {
-                val offset = Calendar.SATURDAY - today
-                calendar.add(Calendar.DATE, offset)
-            }
-            searchTimeViewModel.saveWeekendDate(EventUtils.getSimpleFormattedDate(calendar.time))
-            calendar.add(Calendar.DATE, 1)
-            searchTimeViewModel.saveNextToWeekendDate(EventUtils.getSimpleFormattedDate(calendar.time))
-            searchTimeViewModel.saveDate(THIS_WEEKEND)
-            redirectToSearch()
+            favoriteEventViewModel.searchWeekend(calendar, searchTimeViewModel, activity)
         }
         rootView.searchNextMonth.setOnClickListener {
-            val today = calendar.get(Calendar.DAY_OF_MONTH)
-            val offset = 30 - today
-            calendar.add(Calendar.DATE, offset)
-            searchTimeViewModel.saveNextMonth(EventUtils.getSimpleFormattedDate(calendar.time))
-            calendar.add(Calendar.MONTH, 1)
-            searchTimeViewModel.saveNextToNextMonth(EventUtils.getSimpleFormattedDate(calendar.time))
-            searchTimeViewModel.saveDate(NEXT_MONTH)
-            redirectToSearch()
+            favoriteEventViewModel.searchNextMonth(calendar, searchTimeViewModel, activity)
         }
 
         favoriteEventsRecyclerAdapter.setListener(recyclerViewClickListener)
@@ -151,19 +128,11 @@ class FavoriteFragment : Fragment() {
         return rootView
     }
 
-    private fun redirectToSearch() {
-        val intent = Intent(activity, MainActivity::class.java).setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-        val bundle = Bundle()
-        bundle.putBoolean(TO_SEARCH, true)
-        intent.putExtras(bundle)
-        startActivity(intent)
-    }
-
     private fun showEmptyMessage(itemCount: Int) {
         noLikedText.visibility = if (itemCount == 0) View.VISIBLE else View.GONE
     }
 
-    private fun setCurrentDate(calendar: Calendar) {
+    fun setCurrentDate(calendar: Calendar) {
         val dayLongName = calendar.getDisplayName(Calendar.DAY_OF_WEEK, Calendar.SHORT, Locale.getDefault())
         val month = calendar.getDisplayName(Calendar.MONTH, Calendar.SHORT, Locale.getDefault())
         val date = calendar.get(Calendar.DAY_OF_MONTH)
